// src/models/IWeatherForecastData.ts
// Generated by https://quicktype.io

export interface IWeatherForecastData {
  cod: string;
  message: number;
  cnt: number;
  list: IWeatherForecast[];
  city: City;
}

interface City {
  id: number;
  name: string;
  coord: Coord;
  country: string;
}

interface Coord {
  lat: number;
  lon: number;
}

export interface IWeatherForecast {
  dt: number;
  main: MainClass;
  weather: Weather[];
  clouds: Clouds;
  wind: Wind;
  sys: Sys;
  dt_txt: string;
  rain?: Rain;
  snow?: Rain;
}

interface Clouds {
  all: number;
}

interface MainClass {
  temp: number;
  temp_min: number;
  temp_max: number;
  pressure: number;
  sea_level: number;
  grnd_level: number;
  humidity: number;
  temp_kf: number;
}

interface Rain {
  '3h'?: number;
}

interface Sys {
  pod: string;
}

interface Weather {
  id: number;
  main: string;
  description: string;
  icon: string;
}

interface Wind {
  speed: number;
  deg: number;
}